typo fixes/rewording & italicizing text
Some phrases were appearing a bit wrong to me, so I changed them to what I consider "better".
I also fixed some typos, and italicized the text on the textarea as it was originally supposed to be italic on the English version, as well as correcting it to what it's supposed to be on the Turkish version.